refactor(aicore/filtering): address review comments #11-13
- README: drop the 'Content Filtering & Prompt Shield' bullet from
  Key Features. It's a sub-feature of 'AI Core Integration' already
  on the list; other modules don't fan out their sub-features either.
- Move set_filtering / disable_filtering bodies from filtering/__init__.py
  to a new filtering/_api.py. filtering/__init__.py is now a thin
  re-export surface; logic lives in named modules.
- Inline Severity into _filters.py (where its only consumer
  AzureContentFilter lives) and delete _models.py — a one-enum file
  with a misleading 'models' name. TestSeverity moves to test_filters.py.

749 unit tests pass; flat 13-name public API at sap_cloud_sdk.aicore
is unchanged for external callers.
